refractoring of contrast checker methods
